requirements.txt
absl-py==1.4.0
adapter-transformers==3.2.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
cachetools==5.3.0
datasets==2.10.1
dill==0.3.6
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
importlib-metadata==6.1.0
Markdown==3.4.1
MarkupSafe==2.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.14
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0

requests==2.28.2
numpy==1.22.4
timm==0.6.13